refactor: major rework of API.js and its function
API.js now serves as a middleman connecting to all fspop classes and utils.

Before API.js had its own helper functions to interface with all of fspop's classes.
This was a bad idea as, instead of providing a set of useful abstractions, it limited
the overall functionality with no real gain.